requirements.txt
Django==5.0.1
djangorestframework==3.14.0
weasyprint==63.1
#python-magic==0.4.27
django-cors-headers==4.3.1
whitenoise==6.6.0
html2image==2.0.5
django-health-check==3.18.3
psutil==5.9.8